factor out common driver methods
